transcribe/regulation_checker/pyproject.toml
[project]
name = "regulation_checker"
version = "0.1.0"
description = "regulation_checker using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "agentops>=0.4.18",
    "chromadb>=0.5.23",
    "crewai-tools>=0.55.0",
    "crewai[tools]>=0.148.0,<1.0.0",
    "google-generativeai>=0.8.5",
    "langtrace-python-sdk>=3.8.18",
    "mlflow>=3.1.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "tqdm>=4.67.1",
]

[project.scripts]
regulation_checker = "regulation_checker.main:run"
run_crew = "regulation_checker.main:run"
train = "regulation_checker.main:train"
replay = "regulation_checker.main:replay"
test = "regulation_checker.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
